pyproject.toml
[tool.poetry]
name = "devjam"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django = "^5.0.2"
django-cors-headers = "^4.3.1"
pillow = "^10.2.0"
configparser = "^6.0.1"
mysqlclient = "^2.2.4"
redis = "^5.0.1"
django-redis = "^5.4.0"
django-prometheus = "^2.3.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
